rename dnd-container -> dnd-area
