+++ src/internal/components/dnd-area/__tests__/use-live-announcements.test.tsx
