+++ src/internal/components/dnd-area/keyboard-sensor/defaults.ts
