+++ src/internal/components/dnd-area/keyboard-sensor/index.ts
