+++ src/internal/components/dnd-area/keyboard-sensor/utilities/__tests__/scroll.test.ts
