+++ src/internal/components/dnd-area/keyboard-sensor/utilities/events.ts
