+++ src/internal/components/dnd-area/keyboard-sensor/utilities/listeners.ts
