+++ src/internal/components/dnd-area/keyboard-sensor/utilities/scroll.ts
