+++ src/internal/components/dnd-area/styles.scss
